backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
qdrant-client==1.9.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
openai==1.30.0
redis==5.0.4
beautifulsoup4==4.12.3
requests==2.31.0
pydantic==2.6.4
python-dotenv==1.0.1
httpx==0.27.0
